update the tasklist
