+++ src/Components/index.ts
@@ -1,4 +1,3 @@
 export * from './TaskCard';
 export * from './TaskList';
 export * from './AddTask';
-export * from './DeleteTask';
